+++ src/index.ts
@@ -1,2 +1,1 @@
-export * from './dialecte'
-export * from './config'
+export * from './v1/dialecte'