+++ src/v1/config/namespaces.ts
@@ -1,0 +1,6 @@
+import { XSI_NAMESPACE } from '@dialecte/core/helpers'
+
+export const PLC_NAMESPACES = {
+	default: { uri: 'www.iec.ch/public/TC65SC65BWG7TF10', prefix: '' },
+	xsi: XSI_NAMESPACE,
+} as const